src/API/requirements.txt
fastapi
uvicorn
boto3
pandas
python-dateutil
